Update document.py
document.getHtmlSents() has been revised to handle inline formatting tags (e.g., <strong>, <span> etc.). These tags are no longer included in the HTML strings returned from this method.

This method is used in ontopic.py, line 152

The data returned from getSentStructureData() has been updated to indicate whether or not a sentence is a heading. If it is a heading, we will display the rectangles/squares in gray ("#cccccc").

This method is used in ontopic.py Line 149,