Update document.py
The analysis of lexical overlap for the term matrix visualization has been moved from processDoc(...) to generateGlobalVisData(...) since it requires a lot of time (~10+ seconds) when the input file is large (> 5000 words). This analysis is done only if generateGlobalVisData is called. This should speed up the time to process loadFromHtmlString(...).