Fixed the bug associated with image tags surrounded by other tags, e.g., <strong>.
In the previous version, images included in a Word file were not displayed correctly when inline images are styled with tags like <strong>. This version fixes this problem.

This version also resizes images if they are greater than 400x400px. We may instead want to do this using the width and height attributes associated with the <img> tag.